Move everything over to lib
